pyproject.toml
[tool.poetry]
name = "alpha-store"
version = "0.1.0"
description = ""
authors = ["Nikita Smykov <132088678+Apicqq@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.1"
djangorestframework = "^3.15.2"
pillow = "^10.4.0"
djoser = "^2.2.3"
black = "^24.8.0"
drf-yasg = "^1.21.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
